メイン画面のメニューを切り出す
